requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
cffi==1.14.1              # via cryptography
chardet==3.0.4            # via requests
cryptography==3.0         # via social-auth-core
defusedxml==0.6.0         # via python3-openid, social-auth-core
deprecated==1.2.10        # via pygithub
django-environ==0.4.5     # via -r requirements.in
django==3.1               # via -r requirements.in
idna==2.10                # via requests
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
pycparser==2.20           # via cffi
pygithub==1.52            # via -r requirements.in
pyjwt==1.7.1              # via pygithub, social-auth-core
python3-openid==3.2.0     # via social-auth-core
pytz==2020.1              # via django
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.24.0          # via pygithub, requests-oauthlib, social-auth-core
six==1.15.0               # via cryptography, social-auth-app-django, social-auth-core
social-auth-app-django==4.0.0  # via -r requirements.in
social-auth-core==3.3.3   # via social-auth-app-django
sqlparse==0.3.1           # via django
urllib3==1.25.10          # via requests
wrapt==1.12.1             # via deprecated
